chore: cleanup files from old toolbar
