+++ src/components/App.js
@@ -56,7 +56,6 @@ import { default as TitleBar } from './TitleBar/TitleBar.js'
 import { Toolbar } from './Toolbar/Toolbar.js'
 import StartScreen from './Visualization/StartScreen.js'
 import { Visualization } from './Visualization/Visualization.js'
-import { ToolbarV2 } from './ToolbarV2/ToolbarV2.js'
 
 const visualizationQuery = {
     eventVisualization: {
@@ -311,7 +310,6 @@ const App = () => {
             )}
         >
             <Toolbar onFileMenuAction={onFileMenuAction} />
-            <ToolbarV2 onFileMenuAction={onFileMenuAction} />
             <div
                 className={cx(
                     classes.sectionMain,
