+++ src/components/DownloadMenu/PlainDataSourceSubMenu.js
@@ -2,6 +2,7 @@ import i18n from '@dhis2/d2-i18n'
 import { MenuItem, MenuSectionHeader } from '@dhis2/ui'
 import PropTypes from 'prop-types'
 import React from 'react'
+import { HoverMenuBar } from '../../analyticsComponents/index.js'
 import {
     DOWNLOAD_TYPE_PLAIN,
     ID_SCHEME_UID,
@@ -11,33 +12,44 @@ import {
 
 export const PlainDataSourceSubMenu = ({
     download,
+    hoverable,
     format,
     label,
     ...menuItemProps
-}) => (
-    <MenuItem label={label} {...menuItemProps}>
-        <MenuSectionHeader label={i18n.t('Metadata ID scheme')} />
-        <MenuItem
-            label={i18n.t('ID')}
-            onClick={() => download(DOWNLOAD_TYPE_PLAIN, format, ID_SCHEME_UID)}
-        />
-        <MenuItem
-            label={i18n.t('Code')}
-            onClick={() =>
-                download(DOWNLOAD_TYPE_PLAIN, format, ID_SCHEME_CODE)
-            }
-        />
-        <MenuItem
-            label={i18n.t('Name')}
-            onClick={() =>
-                download(DOWNLOAD_TYPE_PLAIN, format, ID_SCHEME_NAME)
-            }
-        />
-    </MenuItem>
-)
+}) => {
+    const MenuItemComponent = hoverable ? HoverMenuBar.MenuItem : MenuItem
+
+    return (
+        <MenuItemComponent label={label} {...menuItemProps}>
+            <MenuSectionHeader
+                label={i18n.t('Metadata ID scheme')}
+                hideDivider
+            />
+            <MenuItemComponent
+                label={i18n.t('ID')}
+                onClick={() =>
+                    download(DOWNLOAD_TYPE_PLAIN, format, ID_SCHEME_UID)
+                }
+            />
+            <MenuItemComponent
+                label={i18n.t('Code')}
+                onClick={() =>
+                    download(DOWNLOAD_TYPE_PLAIN, format, ID_SCHEME_CODE)
+                }
+            />
+            <MenuItemComponent
+                label={i18n.t('Name')}
+                onClick={() =>
+                    download(DOWNLOAD_TYPE_PLAIN, format, ID_SCHEME_NAME)
+                }
+            />
+        </MenuItemComponent>
+    )
+}
 
 PlainDataSourceSubMenu.propTypes = {
     download: PropTypes.func.isRequired,
     format: PropTypes.string.isRequired,
     label: PropTypes.string.isRequired,
+    hoverable: PropTypes.bool,
 }
