--- src/components/DownloadMenuV2/ModalDownloadDropdown.module.css
@@ -1,4 +1,0 @@
-.container {
-    margin-top: var(--spacers-dp12);
-    margin-bottom: var(--spacers-dp16);
-}
