--- src/components/DownloadMenuV2/constants.js
@@ -1,10 +1,0 @@
-export const DOWNLOAD_TYPE_PLAIN = 'plain'
-export const DOWNLOAD_TYPE_TABLE = 'table'
-export const FILE_FORMAT_CSV = 'csv'
-export const FILE_FORMAT_HTML_CSS = 'html+css'
-export const FILE_FORMAT_JSON = 'json'
-export const FILE_FORMAT_XLS = 'xls'
-export const FILE_FORMAT_XML = 'xml'
-export const ID_SCHEME_UID = 'UID'
-export const ID_SCHEME_CODE = 'CODE'
-export const ID_SCHEME_NAME = 'NAME'
