--- src/components/DownloadMenuV2/index.js
@@ -1,2 +1,0 @@
-export { ToolbarDownloadDropdown } from './ToolbarDownloadDropdown.js'
-export { ModalDownloadDropdown } from './ModalDownloadDropdown.js'
