--- src/components/Toolbar/MenuBar/styles/InterpretationsButton.module.css
@@ -1,4 +1,0 @@
-.iconWrapper {
-    display: flex;
-    margin-right: var(--spacers-dp8);
-}
