--- src/components/Toolbar/MenuBar/styles/MenuBar.module.css
@@ -1,13 +1,0 @@
-.menuBar {
-    background: var(--colors-white);
-    display: flex;
-    align-items: center;
-    padding: 0 var(--spacers-dp8);
-    height: 38px;
-}
-.updateButton {
-    margin-right: var(--spacers-dp8);
-}
-.flexGrow {
-    flex-grow: 1;
-}
