--- src/components/Toolbar/MenuBar/styles/MenuButton.module.css
@@ -1,29 +1,0 @@
-.menuButton {
-    display: inline-flex;
-    position: relative;
-    align-items: center;
-    justify-content: center;
-    font-size: 14px;
-    font-weight: 400;
-    text-transform: none;
-    padding: 6px var(--spacers-dp12);
-    border-radius: 3px;
-    color: var(--colors-grey900);
-    box-sizing: border-box;
-    line-height: 1.75;
-    background: none;
-    border: none;
-    transition: background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
-    cursor: pointer;
-}
-
-.menuButton:hover:enabled {
-    background-color: var(--colors-grey200);
-}
-
-.menuButton:disabled {
-    color: var(--colors-grey400);
-}
-.menuButton:focus {
-    outline: none;
-}
