--- src/components/Toolbar/VisualizationTypeSelector/styles/VisualizationTypeSelector.module.css
@@ -1,25 +1,0 @@
-.container {
-    width: 260px;
-    border-right: 1px solid var(--colors-grey400);
-}
-
-.container.hidden {
-    display: none;
-}
-
-.button {
-    height: 32px;
-    padding: 4px var(--spacers-dp8);
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    background-color: var(--colors-white);
-}
-
-.button > svg {
-    margin-right: var(--spacers-dp8);
-}
-
-.button > span {
-    font-size: 14px;
-}
